Check role and redirect login

+++ controllers/nationController.js
@@ -1,8 +1,9 @@
 const Nations = require("../model/nation");
+const Players = require("../model/player");
 class NationController {
   index(req, res, next) {
     Nations.find({})
-      .then((nations) => {   
+      .then((nations) => {
         res.render("nationSite", {
           title: "The list of Nations",
           nations: nations,
@@ -47,8 +48,21 @@ class NationController {
       });
   }
   delete(req, res, next) {
-    Nations.findByIdAndDelete({ _id: req.params.nationId })
-      .then(() => res.redirect("/nations"))
+    Players.find({ nation: req.params.nationId })
+      .populate("nation")
+      .then((data) => {
+        if (data.length > 0) {
+          req.flash(
+            "error_msg",
+            `You can not delete this nation because it has already been connected with other players`
+          );
+          return res.redirect("/nations");
+        } else {
+          Nations.findByIdAndDelete({ _id: req.params.nationId })
+            .then(() => res.redirect("/nations"))
+            .catch(next);
+        }
+      })
       .catch(next);
   }
 }
